+++ app/build.gradle
@@ -39,4 +39,8 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
 
     implementation 'com.robotemi:sdk:1.134.1'
+    implementation 'androidx.compose.ui:ui-graphics-android:1.7.3'
+    implementation 'androidx.compose.material3:material3-android:1.3.0'
+    implementation 'androidx.navigation:navigation-compose:2.8.2'
+    implementation 'androidx.navigation:navigation-runtime-ktx:2.8.2'
 }
