+++ app/src/main/java/de/fhkiel/temi/robogguide/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.example.rob_temi_robo_ui.ui.theme
+package de.fhkiel.temi.robogguide.ui.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
