+++ app/src/main/java/de/fhkiel/temi/robogguide/ui/theme/components/Header.kt
@@ -1,4 +1,4 @@
-package com.example.rob_temi_robo_ui.ui.theme.components
+package de.fhkiel.temi.robogguide.ui.theme.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.material3.Text
